+++ client/src/styles/profileStyles.js
@@ -4,44 +4,6 @@ export const profileStyles = theme => ({
     root: {
         display: "flex"
     },
-
-    subtitle: {
-        fontWeight: "600",
-        transition: theme.transitions.create("font-size", {
-            easing: theme.transitions.easing.easeInOut,
-            duration: theme.transitions.duration.enteringScreen
-        })
-    },
-    collapsed: {
-        fontSize: "0.7rem"
-    },
-    drawerPaper: {
-        position: "relative",
-        whiteSpace: "nowrap",
-        width: "240px",
-        transition: theme.transitions.create("width", {
-            easing: theme.transitions.easing.sharp,
-            duration: theme.transitions.duration.enteringScreen
-        })
-    },
-    drawerPaperClose: {
-        overflowX: "hidden",
-        transition: theme.transitions.create("width", {
-            easing: theme.transitions.easing.sharp,
-            duration: theme.transitions.duration.leavingScreen
-        }),
-        width: theme.spacing(7),
-        [theme.breakpoints.up("sm")]: {
-            width: theme.spacing(9)
-        }
-    },
-    toolbarIcon: {
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "flex-end",
-        padding: "0 8px",
-        ...theme.mixins.toolbar
-    },
     appBarSpacer: theme.mixins.toolbar,
     content: {
         backgroundColor: theme.palette.gray,
